+++ examples/features/src/lib.rs
@@ -8,6 +8,7 @@ pub mod big_compute_buffers;
 pub mod boids;
 pub mod bunnymark;
 pub mod conservative_raster;
+pub mod cooperative_matrix;
 pub mod cube;
 pub mod hello_synchronization;
 pub mod hello_triangle;